+++ Semana22/src/index.ts
@@ -1,19 +1,7 @@
-import express, {Express} from 'express'
-import cors from 'cors'
-import { AddressInfo } from "net"
+import { app } from './app'
+import getAllPokemons from './endpoints/getAllPokemons'
 
-const app: Express = express();
+//app.post('/pokemon/post', createPokemon)
+app.get ('/pokemon', getAllPokemons)
+//app.get('/pokemon/:id', getPokemonByID)
 
-app.use(express.json());
-app.use(cors());
-
-
-
-const server = app.listen(process.env.PORT || 3003, () => {
-    if (server) {
-       const address = server.address() as AddressInfo;
-       console.log(`Server is running in http://localhost: ${address.port}`);
-    } else {
-       console.error(`Failure upon starting server.`);
-    }
-});
